Refactor bot into modular handlers
