+++ handlers/admin.py
@@ -1,27 +1,65 @@
 import logging
+from contextlib import suppress
 from pyrogram import Client, filters
+from pyrogram.types import Message, ChatPermissions
 from pyrogram.enums import ParseMode
-from pyrogram.types import Message
 
-from utils.db import (
-    approve_user, unapprove_user, get_approved,
-    is_approved, get_approval_mode, toggle_approval_mode,
-    set_approval_mode
-)
 from utils.perms import is_admin
 from utils.errors import catch_errors
+from utils.db import (
+    approve_user,
+    unapprove_user,
+    get_approved,
+    is_approved,
+    toggle_approval_mode,
+    set_approval_mode,
+    get_approval_mode,
+)
 
 logger = logging.getLogger(__name__)
 
 
-def user_mention(user) -> str:
-    """Return HTML-formatted user mention."""
-    return f"<a href='tg://user?id={user.id}'>{user.first_name}</a>"
-
-
 def register(app: Client) -> None:
-    async def require_admin_reply(message: Message, action: str) -> tuple[int, str] | None:
-        """Ensure message is a reply by an admin to a user."""
+    async def admin_action(message: Message, action: str) -> None:
+        if message.chat.type not in {"group", "supergroup"}:
+            await message.reply_text("Group only command", parse_mode=ParseMode.HTML)
+            return
+        if not await is_admin(app, message):
+            await message.reply_text("Admins only", parse_mode=ParseMode.HTML)
+            return
+        if not message.reply_to_message or not message.reply_to_message.from_user:
+            await message.reply_text("Reply to a user's message.")
+            return
+
+        user = message.reply_to_message.from_user
+        try:
+            if action == "ban":
+                await app.ban_chat_member(message.chat.id, user.id)
+            elif action == "kick":
+                await app.ban_chat_member(message.chat.id, user.id)
+                await app.unban_chat_member(message.chat.id, user.id)
+            elif action == "mute":
+                await app.restrict_chat_member(message.chat.id, user.id, ChatPermissions())
+            await message.reply_text(f"{action.title()} successful ✅")
+        except Exception as e:  # noqa: BLE001
+            await message.reply_text(f"❌ Failed: {e}")
+
+    @app.on_message(filters.command("ban"))
+    @catch_errors
+    async def cmd_ban(client: Client, message: Message) -> None:
+        await admin_action(message, "ban")
+
+    @app.on_message(filters.command("kick"))
+    @catch_errors
+    async def cmd_kick(client: Client, message: Message) -> None:
+        await admin_action(message, "kick")
+
+    @app.on_message(filters.command("mute"))
+    @catch_errors
+    async def cmd_mute(client: Client, message: Message) -> None:
+        await admin_action(message, "mute")
+
+    async def require_admin_reply(message: Message, action: str):
         if not await is_admin(app, message):
             await message.reply_text("🚫 <b>Only admins can do this.</b>", parse_mode=ParseMode.HTML)
             return None
@@ -29,13 +67,14 @@ def register(app: Client) -> None:
             await message.reply_text(f"📌 <b>Reply to a user's message to {action}.</b>", parse_mode=ParseMode.HTML)
             return None
         user = message.reply_to_message.from_user
-        return user.id, user_mention(user)
+        return user.id, f"<a href='tg://user?id={user.id}'>{user.first_name}</a>"
 
     @app.on_message(filters.command("approve") & filters.group)
     @catch_errors
     async def approve_cmd(client: Client, message: Message):
         result = await require_admin_reply(message, "approve")
-        if result is None: return
+        if result is None:
+            return
         user_id, mention = result
         await approve_user(message.chat.id, user_id)
         await message.reply_text(f"✅ <b>Approved</b> {mention}", parse_mode=ParseMode.HTML)
@@ -44,7 +83,8 @@ def register(app: Client) -> None:
     @catch_errors
     async def unapprove_cmd(client: Client, message: Message):
         result = await require_admin_reply(message, "unapprove")
-        if result is None: return
+        if result is None:
+            return
         user_id, mention = result
         await unapprove_user(message.chat.id, user_id)
         await message.reply_text(f"❌ <b>Unapproved</b> {mention}", parse_mode=ParseMode.HTML)
@@ -55,15 +95,11 @@ def register(app: Client) -> None:
         if not await is_admin(app, message):
             await message.reply_text("🚫 <b>Only admins can view approvals.</b>", parse_mode=ParseMode.HTML)
             return
-
         users = await get_approved(message.chat.id)
         if not users:
             await message.reply_text("📭 <i>No approved users found.</i>", parse_mode=ParseMode.HTML)
             return
-
-        text = "<b>📋 Approved Users:</b>\n"
-        for uid in users:
-            text += f"• <code>{uid}</code>\n"
+        text = "<b>📋 Approved Users:</b>\n" + "\n".join(f"• <code>{u}</code>" for u in users)
         await message.reply_text(text, parse_mode=ParseMode.HTML)
 
     @app.on_message(filters.command("approval") & filters.group)
@@ -72,8 +108,6 @@ def register(app: Client) -> None:
         if not await is_admin(app, message):
             await message.reply_text("🔒 <b>You must be an admin to change approval mode.</b>", parse_mode=ParseMode.HTML)
             return
-
-        # Toggle or set approval mode based on command
         if len(message.command) == 1:
             enabled = await toggle_approval_mode(message.chat.id)
         else:
@@ -87,32 +121,7 @@ def register(app: Client) -> None:
             else:
                 await message.reply_text("❗ <b>Usage:</b> <code>/approval [on|off]</code>", parse_mode=ParseMode.HTML)
                 return
-
         await message.reply_text(
             f"🔄 <b>Approval mode is now {'ENABLED ✅' if enabled else 'DISABLED ❌'}</b>",
-            parse_mode=ParseMode.HTML
+            parse_mode=ParseMode.HTML,
         )
-
-    @app.on_message(filters.group, group=10)
-    @catch_errors
-    async def enforce_approval(client: Client, message: Message):
-        user = message.from_user
-        chat_id = message.chat.id
-
-        if not user or user.is_bot:
-            return
-
-        if await is_admin(client, message, user.id):
-            return
-
-        if not await get_approval_mode(chat_id):
-            return
-
-        if await is_approved(chat_id, user.id):
-            return
-
-        try:
-            await message.delete()
-            logger.info("Deleted message from unapproved user %s in chat %s", user.id, chat_id)
-        except Exception as e:
-            logger.warning("Failed to delete message from %s: %s", user.id, str(e))
